requirements.txt
streamlit==1.40.2
pinecone-client==5.0.1
python-dotenv==1.0.1
opencv-python==4.10.0.84
pytesseract==0.3.13
pillow==11.0.0
transformers==4.46.3
torch==2.2.2
openai==1.55.3
numpy==1.26.4
sentence-transformers==3.3.1
groq==0.12.0